# gamification-impl/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Gamification API
  description: An API to demonstrate Swagger and Spring Boot
security:
  #  - bearerAuth: []
  - apiKeyAuth: [ ]
paths:
  /applications:
    post:
      description: register an application
      operationId: registerApplication
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/ApplicationRegistration'
      responses:
        '201':
          description: created
            #get:
            # description: get the list of all applications
            #operationId: getApplications
            #responses:
            #  '200':
            #   description: success
          #  content:
          #   'application/json':
          #    schema:
          #     type: array
          #    items:
          #     $ref: '#/components/schemas/Application'
  /badges:
    post:
      description: create a badge
      operationId: createBadge
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Badge'
      responses:
        '201':
          description: created
    get:
      description: get the list of all badges
      operationId: getBadges
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Badge'
  /badges/{id}:
    get:
      description: get one badge, by id
      operationId: getBadge
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
  /events:
    post:
      description: create an event
      operationId: createEvent
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: created
  /users:
    get:
      description: get the list of all users
      operationId: getUsers
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      description: get one user, by id
      operationId: getUser
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/User'
  /rules:
    post:
      description: create a rule
      operationId: createRule
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Rule'
      responses:
        '201':
          description: created
components:
  securitySchemes:
    #    bearerAuth:            # arbitrary name for the security scheme
    #      type: http
    #      scheme: bearer       # warning: it HAS TO be a lowercase b for the swagger ui version used in springfox
    #      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    Application:
      type: object
      properties:
        name:
          type: string
        key:
          type: string
    ApplicationRegistration:
      type: object
      properties:
        name:
          type: string
    Badge:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    Event:
      type: object
      properties:
        userId:
          type: string
        timestamp:
          type: string
          format: date-time
        eventType:
          type: string
        properties:
          type: object
          properties:
            subType:
              type: string
            quantity:
              type: integer
    User:
      type: object
      properties:
        userId:
          type: string
        nbBadges:
          type: integer

    Rule:
      type: object
      properties:
        name:
          type : string
        if:
          type: object
          properties:
            ruleType:
              type: string
        then:
          type: object
          properties:
            awardBadge:
              type: string
            awardPoints:
              type: object
              properties:
                pointScale:
                  type: string
                amount:
                  type: integer